pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DecoraterBotUtils"
version = "0.0.2"
license = { text="MIT" }
authors = [
  { name="Decorater", email="seandhunt_7@yahoo.com" },
]
description = "This package is for bringing various things to DecoraterBot."
# readme = "README.rst"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    'regex',
    'colorama',
    'discord.py',
    'aiohttp',
    'aiofiles'
]

[project.urls]
"Homepage" = "https://github.com/DecoraterBot-devs/DecoraterBotUtils"
"Bug Tracker" = "https://github.com/DecoraterBot-devs/DecoraterBotUtils/issues"
